project/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
redis>=4.1.0
python-multipart>=0.0.5
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
